refactor(client): react hot toast implementation
