+++ client/src/utils/index.ts
@@ -1,4 +1,8 @@
+export * from './api';
 export * from './convertStringToHTML';
 export * from './errorHandlers';
 export * from './formatDate';
-export * from './toastConfig';
+export * from './reactHotToast';
+
+// export client components
+export { default as Toaster } from './reactHotToast/Toaster';
