+++ client/src/utils/reactHotToast/Toaster.tsx
@@ -1,0 +1,5 @@
+'use client';
+
+import { Toaster } from 'react-hot-toast';
+
+export default Toaster;
